# apps/ai-orchestrator/openapi/orchestrator-openapi.yaml
openapi: 3.1.0
info:
  title: AI Orchestrator API
  version: 0.1.0
  description: >
    Minimal orchestration API for generating YAML test cases via the test-design-agent
    and optionally dispatching the web-playwright-python runner.
servers:
  - url: http://127.0.0.1:8000
paths:
...
  /assets/page-objects:
    post:
      summary: Create a page object asset
      operationId: postPageObject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreatePageObjectRequest"
      responses:
        "201":
          description: Page object created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PageObjectAssetResponse"
        "422":
          description: Request validation failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Internal asset management failure
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
          properties:
            code:
              type: string
              enum:
                - invalid_json
                - not_found
                - unsupported_media_type
                - validation_error
                - runner_failed
                - internal_error
            message:
              type: string
            details:
              type: object
              additionalProperties: true
...
    CreatePageObjectRequest:
      type: object
      additionalProperties: false
      required:
        - page
      properties:
        page:
          type: string
          minLength: 1
        description:
          type: string
...
    PageObjectAssetResponse:
      type: object
      additionalProperties: false
      required:
        - page_object
        - path
      properties:
        page_object:
          type: object
        path:
          type: string
...